# Layer Panel: replace inactive text Input with Div element