--- a/src/style/component-styles.less
+++ b/src/style/component-styles.less
@@ -21,7 +21,6 @@
 @import (multiple) "./shared/scrollbar.less";
 @import (multiple) "./shared/range.less";
 @import (multiple) "./shared/title-header.less";
-@import (multiple) "./shared/text-input.less";
 @import (multiple) "./shared/select.less";
 @import (multiple) "./shared/split-button.less";
 @import (multiple) "./shared/link-list.less";