… src/style/shared/text-input.less