Chrome has a bug when rendering inactive (disabled) text input elements, which results in increased rendering time. This PR changes the `TextInput` element to work around the issue: instaed of rendering text input with `disabled` attribute, it renders a div element when the component is not in edit mode. The performance improvement is significant, and below is my test result (the numbers are **brower rendering time**):

  | Before   | After | Change
------------- | -------------| ----- | ----
Layer selection | 32ms      | **21ms** | -11ms
AB selection    | 286ms     | **29ms** | -257ms
Collapse AB     | **37ms**      | 48ms | +11ms
Expand AB       | 326ms     | **91ms** | -235ms
AB Visibility   | 309ms     | **82ms** | -227ms

Below are the panels with style issues due to the change, and should be addressed in a separate PR.
* libraries panel
* export panel

**Credits to @iwehrman and @designedbyscience for their initial research/experiments**