@iwehrman 2016-01-18
I don't think click events work in `LayerFace` when it's in editing mode. I can double-click to enter editing mode, but then I can't click to position the cursor.