@iwehrman 2016-01-18
This also drastically reduces the time to switch into Vermilion from an untitled document, from ~1200ms to 400ms.