@iwehrman 2016-01-18
The descenders are also clipped in `LayerFace`:
![image](https://cloud.githubusercontent.com/assets/1075154/12405654/f1f51266-bdfa-11e5-988d-f08f1cc3ac3a.png)

This is something that I'm pretty sure I explicitly fixed in my original branch.