@shaoshing 2016-01-19
@iwehrman I did not see the style bug your mentioned in different screen resolution. Can you show me your screen configuration? I found the `grunt debug` task sometimes did not pick up the change in LESS files unless I restart it, may be this is the cause? 

<img width="459" alt="screen shot 2016-01-18 at 4 41 35 pm" src="https://cloud.githubusercontent.com/assets/118264/12406903/58bd8054-be06-11e5-9f69-6bfb94fb7602.png">
![screen shot 2016-01-18 at 4 57 41 pm](https://cloud.githubusercontent.com/assets/118264/12406904/58c0b8be-be06-11e5-82d8-96bd9326709e.png)