@iwehrman 2016-01-19
I only have non-retina to test at the moment, but it looks good at different UI sizes, so I think your theory about `grunt debug` not catching the updated styles are correct. The updated `TextInput` structure looks good, and editing works again. Thanks for pushing this over the line, @shaoshing!